+++ src/main/java/org/alfresco/support/expert/jmxdumpanalyzerfx/JmxAppliedToServerController.java
@@ -5,11 +5,7 @@
  */
 package org.alfresco.support.expert.jmxdumpanalyzerfx;
 
-import java.io.BufferedReader;
-import java.io.FileReader;
 import java.io.IOException;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.net.URL;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -49,9 +45,6 @@ public class JmxAppliedToServerController implements Initializable {
     }
 
 	private void doWork() throws IOException{
-		String line; // applied version info
-	
-		BufferedReader br = new BufferedReader(new FileReader(jmxFilePath)); 
 
 		populateVersionInfo("APPLIED VERSIONS (upgrade path)");
 		populateVersionInfo("");