+++ src/main/java/org/alfresco/support/expert/jmxdumpanalyzerfx/JmxSOLRController.java
@@ -8,12 +8,9 @@ package org.alfresco.support.expert.jmxdumpanalyzerfx;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.net.URL;
 
 import java.util.ResourceBundle;
-import java.util.Set;
 
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
@@ -223,6 +220,7 @@ public class JmxSOLRController implements Initializable {
 		    }			    
  
 		   }
+		br.close();
 }